release 3.1.120 source code

+++ services/kvs/src/main/java/com/huaweicloud/sdk/kvs/v1/model/ScanKvResponse.java
@@ -25,7 +25,7 @@ public class ScanKvResponse extends SdkBsonDocResponse {
     @JsonProperty(value = "returned_count")
     @BsonProperty(value = "returned_count")
 
-    private Integer returnedCount;
+    private Long returnedCount;
 
     @JsonInclude(JsonInclude.Include.NON_NULL)
     @JsonProperty(value = "filtered_count")
@@ -53,7 +53,7 @@ public class ScanKvResponse extends SdkBsonDocResponse {
 
     private static Codec<ScanKvResponse> codec;
 
-    public ScanKvResponse withReturnedCount(Integer returnedCount) {
+    public ScanKvResponse withReturnedCount(Long returnedCount) {
         this.returnedCount = returnedCount;
         return this;
     }
@@ -62,11 +62,11 @@ public class ScanKvResponse extends SdkBsonDocResponse {
      * 返回的文档数量，为0不表示结束。 > 如果filtered_count超过500仍无匹配，则返回0。
      * @return returnedCount
      */
-    public Integer getReturnedCount() {
+    public Long getReturnedCount() {
         return returnedCount;
     }
 
-    public void setReturnedCount(Integer returnedCount) {
+    public void setReturnedCount(Long returnedCount) {
         this.returnedCount = returnedCount;
     }
